patrolswarm/requirements.txt
# Patrol Swarm — Python dependencies
# Install: pip install -r requirements.txt

# LangChain + LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0

# LangGraph checkpointers — durable blackboard persistence
langgraph-checkpoint-sqlite>=1.0.0   # SQLite (dev / single-node)
aiosqlite>=0.20.0                     # async SQLite driver
# For production PostgreSQL, also install:
# langgraph-checkpoint-postgres>=1.0.0
# psycopg[binary]>=3.1.0

# Pydantic v2
pydantic>=2.0.0

# APScheduler
APScheduler>=3.10.0

# Environment variable loading
python-dotenv>=1.0.0

# Async
asyncio-throttle>=1.0.2

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0

# HTTP client — fire-and-forget POST to investigation API
httpx>=0.27.0

# Optional: dev/test
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: evaluation harness (eval/)
# Install with: pip install -r requirements.txt matplotlib numpy datasets tqdm scikit-learn seaborn
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
# datasets>=2.14.0       # HuggingFace — install separately to keep core deps lean
# scikit-learn>=1.3.0    # used by metrics.py if present
# seaborn>=0.12.0        # optional nicer charts
